add a whitespace in the last code section of how-to/using-mongo-db.md
The whitespace in the braces of "const { MongoClient} = req..." was previously unbalanced (having one whitespace to the left, but none to the right). This commit adds a whitespace so the line is changed to "const { MongoClient } = req...".
(I can revert this commit if the previous style is preferred)
I may also point out how the line "if(err){ console.error(err); return false;}" is formatted. (The line is located 17 lines after the aforementioned mongodb import line). This commit does not change that line, but I find the spacing of that line quite odd.